Bot/requirements.txt
discord.py
asyncpg 